% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.rails_fit}
\alias{plot.rails_fit}
\title{Plot the RAILS weight distribution}
\usage{
\method{plot}{rails_fit}(x, breaks = 40, ...)
}
\arguments{
\item{x}{A \code{rails_fit}.}

\item{breaks}{Passed to \code{\link[graphics:hist]{graphics::hist()}}.}

\item{...}{Passed to \code{\link[graphics:hist]{graphics::hist()}}.}
}
\value{
\code{x}, invisibly. Called for the plot.
}
\description{
Draws a histogram of the fitted weights on a log scale, which is where a
heavy right tail -- the usual failure mode of propensity weighting -- is
visible.
}
\examples{
pop <- rails_simulate(2000, seed = 3)
ref <- pop[pop$s == 1, ]; ref$dweight <- 1 / ref$ps
fit <- rails(pop[pop$aou == 1, ], ref, c("agegroup", "sex"),
             weights_ref = "dweight", verbose = FALSE)
plot(fit)

}
