% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAILS-package.R
\docType{package}
\name{RAILS-package}
\alias{RAILS}
\alias{RAILS-package}
\title{RAILS: Raking with Assisted Nested Propensity Score and LIFO Selection}
\description{
Combines a non-probability sample (e.g. a volunteer biobank) with a probability reference sample to produce population-representative weights. The core estimator fits a pseudo-likelihood propensity model on all interactions up to a base order, selects higher-order interactions by a forward likelihood-ratio test, and applies last-in-first-out stepwise raking to reference population margins. All computation runs on aggregated covariate cells. Variance is available from a stacked estimating-equation sandwich that accounts for uncertainty in the estimated weights, and a subgroup wrapper runs the procedure within levels of a stratifying variable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/extrasane/RAILSpkg}
  \item \url{https://extrasane.github.io/RAILSpkg/}
  \item Report bugs at \url{https://github.com/extrasane/RAILSpkg/issues}
}

}
\author{
\strong{Maintainer}: Huiding Chen \email{extrasane@gmail.com} (\href{https://orcid.org/0000-0002-5102-434X}{ORCID})

}
\keyword{internal}
