% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.rails_fit}
\alias{summary.rails_fit}
\alias{print.summary.rails_fit}
\title{Summarize a RAILS fit}
\usage{
\method{summary}{rails_fit}(object, ...)

\method{print}{summary.rails_fit}(x, ...)
}
\arguments{
\item{object}{A \code{rails_fit}.}

\item{...}{Ignored.}

\item{x}{A \code{summary.rails_fit}.}
}
\value{
A \code{summary.rails_fit} object: a list with the selection path, the
terms actually raked to, weight diagnostics from \code{\link[=weight_summary]{weight_summary()}}, the
design effect, and the effective sample size.
}
\description{
Summarize a RAILS fit
}
\examples{
pop <- rails_simulate(2000, seed = 3)
ref <- pop[pop$s == 1, ]; ref$dweight <- 1 / ref$ps
fit <- rails(pop[pop$aou == 1, ], ref, c("agegroup", "sex"),
             weights_ref = "dweight", verbose = FALSE)
summary(fit)

}
